% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{f_adjust_ind_number_1940}
\alias{f_adjust_ind_number_1940}
\title{This function corrects the way some industries appeared in the
Alphabetic Index of Titles. One way of denoting industries is by
enclosing their index in parenthesis (x1). Sometimes a parenthesis
is missing or there are extra spaces}
\usage{
f_adjust_ind_number_1940(x)
}
\arguments{
\item{x}{vector of characters}
}
\value{
Vector of the same size as x but with different substitutions
}
\description{
This function corrects the way some industries appeared in the
Alphabetic Index of Titles. One way of denoting industries is by
enclosing their index in parenthesis (x1). Sometimes a parenthesis
is missing or there are extra spaces
}
