% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{f_change_words}
\alias{f_change_words}
\title{This function homogenizes some symbols that appeared in the
Alphabetic Index of Titles}
\usage{
f_change_words(x)
}
\arguments{
\item{x}{vector of characters}
}
\value{
Vector of the same size as x but with different substitutions
}
\description{
This function homogenizes some symbols that appeared in the
Alphabetic Index of Titles
}
