% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{f_clean_conj}
\alias{f_clean_conj}
\title{This function erases some conjunctions, and corrects some symbols that
appeared in the Alphabetic Index of Titles.}
\usage{
f_clean_conj(x)
}
\arguments{
\item{x}{vector of characters}
}
\value{
Vector of the same size as x but with different substitutions
}
\description{
This function erases some conjunctions, and corrects some symbols that
appeared in the Alphabetic Index of Titles.
}
