% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{f_clean_punct}
\alias{f_clean_punct}
\title{This function corrects some punctuation mistakes that appeared in titles in the
Alphabetic Index of Titles}
\usage{
f_clean_punct(x, flag = 1)
}
\arguments{
\item{x}{vector of characteres}

\item{flag}{Flag to specify if commas should be substituted by space (by default, flag=1 )}
}
\value{
Vector of the same size as x but with different substitutions
}
\description{
This function corrects some punctuation mistakes that appeared in titles in the
Alphabetic Index of Titles
}
