% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{f_clean_words}
\alias{f_clean_words}
\title{This function erases a set of symbols that appear in titles in the
Alphabetic Index of Titles.
This function is a companion to f_change_words.}
\usage{
f_clean_words(x)
}
\arguments{
\item{x}{vector of characteres}
}
\value{
Vector of the same size as x but with different substitutions
}
\description{
This function erases a set of symbols that appear in titles in the
Alphabetic Index of Titles.
This function is a companion to f_change_words.
}
