% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressions.R
\name{f_order_reg}
\alias{f_order_reg}
\title{Function to organize regression coefficients.}
\usage{
f_order_reg(X, Order, order_var)
}
\arguments{
\item{X}{data.table containing the variable names order_var}

\item{order_var}{Variable in the containing the variable names}

\item{order}{List of vectors in the desired order}
}
\description{
Function to organize regression coefficients.
}
