% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressions.R
\name{f_reg_results}
\alias{f_reg_results}
\title{Function to create coefficients with significance stars}
\usage{
f_reg_results(Regresults, arg_round = 4)
}
\arguments{
\item{Regresults}{data.table with 5 columns. (1) Name of variables, (2) Regression coefficient, (3) Standard value, (4) T-Value, (5) P_value}

\item{arg_round}{number of significance digits. 4 by default}
}
\description{
Function to create coefficients with significance stars
}
