% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressions.R
\name{f_regressions}
\alias{f_regressions}
\title{Function to ran multiple regressions}
\usage{
f_regressions(
  arg_formula,
  arg_data,
  arg_weights,
  method,
  arg_clusters,
  arg_fe,
  to_iter = 1000,
  by_iter = 10
)
}
\arguments{
\item{arg_data}{data.table with the variables of the regression}

\item{arg_weights}{variable in arg_data containing the weights used in the regression}

\item{method}{a function either "lm", "lm_robust" or "felm"}

\item{arg_clusters}{variable in arg_data used to cluster standard errors.}

\item{arg_fe}{character with the name of the fe variable. This variable will
only be used if you specify the "felm" method. If you have more than
one fixed effect you should specify it like "fe1 + fe2"}

\item{arg_forlula}{LIST(!!) with the set of regression formulas
(It has to be a list, but can be a list with 1 element).
Formula argument is of the form: y ~ x1 + x2}

\item{iter_vec}{vector containing the numbers of the iteration that you want
to be displayed (default is 10)}
}
\value{
A data.table with the regression coefficients in the different models
}
\description{
Currently you can work with 3 different types of regression algorithms: lm, lm_robust and felm
Currently the package doesn't allow for instrumental variables
}
